pyproject.toml
[tool.poetry]
name = "travel-planner"
version = "0.1.0"
description = ""
authors = ["Ssuwani <jsuwan961205@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.25"
langgraph = "^0.4.7"
streamlit = "^1.45.1"
langchain-openai = "^0.3.18"
python-dotenv = "^1.1.0"
tavily-python = "^0.7.2"
google-api-python-client = "^2.170.0"
google-auth-oauthlib = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
